k3s: make keeping state optional
this commit reverts a behavior in the k3s module introduced in 710cadd605e39f40d0b2512e108e6f399ecb466f
where the temp mount was replaced with a cache volume. This should have
been introduced in a backwards compatible way but it was not. We're
taking care of that now by adding a keepState constructo option. I've
also made other changes so the Config and examples are more robust

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>